refactor shape
