--- src/model/entity/EntityShaper.php
@@ -1,15 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: INFORMATIQUE
- * Date: 07/12/2016
- * Time: 11:12
- */
-
-namespace base\model\entity;
-
-
-interface EntityShaper
-{
-    public function shape(Entity $entity);
-}
